requirements.txt
Django==5.1.2
Pillow==10.4.0
django-filter==24.3